fix code style

+++ codes/cpp/chapter_graph/graph_adjacency_matrix.cpp
@@ -26,7 +26,7 @@ public:
     }
 
     /* 获取顶点数量 */
-    size_t size() const {
+    int size() const {
         return vertices.size();
     }
